Restructure data_processing package
