+++ src/data_processing/video_metadata_download.py
@@ -1,0 +1,161 @@
+import os
+from pytube import (
+    YouTube,
+    Playlist
+)
+import re
+import yt_dlp
+
+
+def download_youtube_video_pytube(url: str, resolution: str="720p"):
+    """
+    Download video from YouTube using the pytube library.
+
+    Args:
+        url (str): URL of the YouTube video.
+        resolution (str, optional): The desired resolution of the video. Defaults to "720p".
+    
+    Raises:
+        Exception: For errors during download.
+
+    Example:
+        download_youtube_video("https://www.youtube.com/watch?v=example", resolution="1080p")
+    """
+    try:
+        save_path = "media/videos/"
+        os.makedirs(save_path, exist_ok=True)
+
+        yt = YouTube(url)
+        stream = yt.streams.filter(res=resolution).first()
+        if stream is None:
+            stream = yt.streams.get_highest_resolution()
+        
+        stream.download(output_path=save_path)
+        print(f"Video download successfully to {save_path}")
+
+    except Exception as e:
+        raise e
+
+
+def download_youtube_video_yt_dlp(url: str):
+    """
+    Download video from YouTube using the pytube library.
+
+    Args:
+        url (str): URL of the YouTube video.
+    
+    Raises:
+        Exception: For errors during download.
+
+    Example:
+        download_youtube_video_yt_dlp("https://www.youtube.com/watch?v=example")
+    """
+    videoid = extract_youtube_video_id(url)
+
+    ydl_opts = {
+            'format': 'best',
+            'outtmpl': f"media/videos/{videoid}.%(ext)s",
+            'retries': 3,
+            'geo_bypass': True,
+        }
+
+    try:
+        with yt_dlp.YoutubeDL(ydl_opts) as ydl:
+            ydl.download([url])
+
+    except Exception as e:
+        raise e
+
+
+def extract_video_urls_from_playlist(url: str):
+    """
+    Extract all YouTube video links from a playlist.
+
+    Args:
+        url (str): URL of the YouTube playlist.
+
+    Returns:
+        list (str): List of video URLs from the playlist.
+    
+    Example: 
+        extract_youtube_video_url_from_playlist("https://www.youtube.com/playlist?list=PLexample")
+    """
+
+    try:
+        playlist = Playlist(url)
+        video_urls = [
+            video_url for video_url in playlist.video_urls
+        ]
+        return video_urls
+    
+    except Exception as e:
+        raise e
+
+
+def extract_meta_data(url: str):
+    """
+    Extract the relevant meta data for a YouTube video.
+
+    Args:
+        url (str): URL of the YouTube video.
+    
+    Raises:
+        Exception: For errors during meta data extraction.
+    """
+    videoid = extract_youtube_video_id(url)
+
+    meta_data = {}
+
+    ydl_opts = {
+            'format': 'best',
+            'outtmpl': f"media/videos/{videoid}.%(ext)s",
+            'retries': 3,
+            'geo_bypass': True,
+        }
+
+    try:
+        with yt_dlp.YoutubeDL(ydl_opts) as ydl:
+            result = ydl.extract_info(url, download=False) # Extract info
+
+            meta_data['id'] = result.get('id')
+            meta_data['title'] = result.get('title')
+            meta_data['description'] = result.get('description')
+            meta_data['upload_date'] = result.get('upload_date')
+            meta_data['duration'] = result.get('duration')
+            meta_data['view_count'] = result.get('view_count')
+            meta_data['uploader_url'] = result.get('uploader_url')
+            meta_data['uploader_id'] = result.get('uploader_id')
+            meta_data['channel_id'] = result.get('channel_id')
+            meta_data['uploader'] = result.get('uploader')
+            meta_data['thumbnail'] = result.get('thumbnail')
+            meta_data['like_count'] = result.get('like_count')
+            meta_data['tags'] = result.get('tags')
+            meta_data['categories'] = result.get('categories')
+            meta_data['age_limit'] = result.get('age_limit')
+
+            return meta_data
+
+    except Exception as e:
+        raise e
+
+
+def extract_youtube_video_id(url: str):
+    """
+    Extract YouTube video id from YouTube video.
+
+    Args:
+        url (str): URL of a YouTube video.
+
+    Returns:
+        video_id (str): ID of the YouTube video.
+
+    Example:
+        extract_youtube_video_id("https://www.youtube.com/playlist?v=example")
+    """
+
+    video_id = re.search(r"[?&]v=([a-zA-Z0-9_-]{11})", url)
+    if video_id:
+        return video_id.group(1)
+    else:
+        raise ValueError("YouTube Video ID could not be extracted from the URL")
+
