+++ src/data_processing/visual_processing.py
@@ -1,0 +1,86 @@
+import os
+import PIL.Image
+import cv2
+from dotenv import load_dotenv
+import google.generativeai as genai
+
+# Static variables
+VIDEO_DIRECTORY = "./media/videos/"
+FRAMES_DIRECTORY = "./media/frames"
+
+# Env variables
+load_dotenv() 
+API_KEY_GOOGLE_GEMINI = os.getenv("API_KEY_GOOGLE_GEMINI")
+
+
+def extract_frames_from_video(video_filepath: str, interval_in_sec: int = 30):
+    """
+    Extract a certain number of frames of a YouTube video.
+
+    Args:
+        video_filepath (str): The filepath of the video.
+        interval_in_sec (int): Interval in seconds between each frame to be extracted. This specifies how frequently frames are captured from the video.
+
+    Example: 
+        extract_youtube_video_id("./media/videos/my_example_video.mp4", 5)
+    """
+
+    print("Begin extraction...")
+
+    filename = video_filepath.replace(VIDEO_DIRECTORY, "")
+    filename = filename.replace(".mp4", "")
+    if not os.path.exists(f"./media/frames/{filename}"):
+        os.makedirs(f"./media/frames/{filename}")
+
+    cam = cv2.VideoCapture(video_filepath)
+    video_fps = cam.get(cv2.CAP_PROP_FPS)
+    interval = round(video_fps * interval_in_sec)
+    print(f"Calculated interval: {interval}")
+
+    success, image = cam.read()
+    count = 0
+    while success:
+        if count % interval == 0:
+            print(f"Writing frame {count}")
+            cv2.imwrite(f"./media/frames/{filename}/frame{count}.jpg", image)
+        success, image = cam.read()
+        count += 1
+
+
+def create_image_description(videoid: str, gemini_model: str="gemini-1.5-flash"):
+    """
+    """
+    genai.configure(api_key=API_KEY_GOOGLE_GEMINI)
+    model = genai.GenerativeModel(gemini_model)
+
+    path_dir = FRAMES_DIRECTORY + "/" + videoid + "/"
+    all_files = os.listdir(path_dir)
+
+    for file in all_files:
+
+        image_file_path = path_dir + "/" + file
+        image_file = PIL.Image.open(image_file_path)
+
+        prompt = (
+            "Please provide a detailed description of the image from the YouTube video, "
+            "focusing on the relevant content, e.g. AI, Machine Learning, "
+            "or other topics shown in the picture. Describe the setting, including the environment, any diagrams, "
+            "charts, equations, or visual aids displayed. If there are visual representations of algorithms, "
+            "models, or data, explain them in detail, including any colors, patterns, or structures. "
+            "If relevant people, such as experts in computer vision or similar fields, are shown in the "
+            "context of the task, briefly mention them and their relevance. Avoid mentioning irrelevant people "
+            "unless they are directly related to the content of the video."
+        )
+
+        response = model.generate_content(
+            [prompt, image_file]
+        )
+
+        path_dir_frame_desc = f"media/frames_description/{videoid}"
+        if not os.path.exists(path_dir_frame_desc):
+            os.makedirs(path_dir_frame_desc)
+
+        filename = file.replace(".jpg", "")
+        with open(f"{path_dir_frame_desc}/{filename}.txt", "w", encoding="utf-8") as datei:
+            datei.write(response.text)
+
